+++ LNKWallet/withdrawpage/withdrawrecordwidget.cpp
@@ -142,31 +142,17 @@ void WithdrawRecordWidget::showWithdrawRecord(QString _accountAddress, QString _
         AssetInfo feeAssetInfo = HXChain::getInstance()->assetInfoMap.value(feeAssetId);
         ui->withdrawRecordTableWidget->setItem(rowCount,3, new QTableWidgetItem(getBigNumberString(feeAmount, feeAssetInfo.precision)));
 
-        int state = -1;
-        if(HXChain::getInstance()->crosschainWithdrawStateMap.contains(transactionId))
+        if(ts.trxState == "withdraw_without_sign_trx_uncreate")
         {
-            state = HXChain::getInstance()->crosschainWithdrawStateMap.value(transactionId);
-        }
-
-        switch (state)
-        {
-        case -1:
-            ui->withdrawRecordTableWidget->setItem(rowCount,5, new QTableWidgetItem(tr("canceled")));
-            break;
-        case 0:
             ui->withdrawRecordTableWidget->setItem(rowCount,5, new QTableWidgetItem(tr("waiting for signature")));
-            break;
-        case 1:
+        }
+        else if( ts.trxState == "withdraw_without_sign_trx_create" || ts.trxState == "withdraw_combine_trx_create")
+        {
             ui->withdrawRecordTableWidget->setItem(rowCount,5, new QTableWidgetItem(tr("being signed")));
-            break;
-        case 3:
-            ui->withdrawRecordTableWidget->setItem(rowCount,5, new QTableWidgetItem(tr("being signed")));
-            break;
-        case 4:
+        }
+        else if( ts.trxState == "withdraw_transaction_confirm")
+        {
             ui->withdrawRecordTableWidget->setItem(rowCount,5, new QTableWidgetItem(tr("confirmed")));
-            break;
-        default:
-            break;
         }
 
         rowCount++;
